+++ CHANGELOG.md
@@ -18,6 +18,12 @@ Recommendation: for ease of reading, use the following order:
 - GQL: `molecule` area: use `BigInt` for `ipnft_token_id` 
 - Allow `molecule` and `molecule.dev` accounts separation
 
+## [0.245.4] - 2025-08-13
+### Added
+- GQL: `dataset_metadata`: Added new method `MetadataProjection` to get `[MetadataExtendedBlock]` filtered by types and range
+### Fixed
+- Return bad request response for requests with empty bearer authentication token
+
 ## [0.245.3] - 2025-07-29
 ### Fixed
 - GQL playground is accessible in anonymous mode